Use unsigned int when dealing with integers that need bitwise operations.  Replace WXDWORD with DWORD (windows). In the future, DWORD types won't be in the interface at all.

+++ demos/forty/forty.cpp
@@ -298,7 +298,7 @@ FortyFrame::ToggleCardSize(wxCommandEvent& event)
 //----------------------------------------------------------------------------
 
 FortyAboutDialog::FortyAboutDialog( wxWindow *parent, wxWindowID id, const wxString &title,
-    const wxPoint &position, const wxSize& size, long style ) :
+    const wxPoint &position, const wxSize& size, unsigned int style ) :
     wxDialog( parent, id, title, position, size, style )
 {
     AddControls(this);
