+++ include/wx/aui/floatpane.h
@@ -37,7 +37,7 @@ public:
                    wxAuiManager* ownerMgr,
                    const wxAuiPaneInfo& pane,
                    wxWindowID id = wxID_ANY,
-                   long style = wxRESIZE_BORDER | wxSYSTEM_MENU | wxCAPTION |
+                   unsigned int style = wxRESIZE_BORDER | wxSYSTEM_MENU | wxCAPTION |
                                 wxFRAME_NO_TASKBAR | wxFRAME_FLOAT_ON_PARENT |
                                 wxCLIP_CHILDREN
                    );
