+++ include/wx/cshelp.h
@@ -72,7 +72,7 @@ public:
                         wxWindowID id = wxID_CONTEXT_HELP,
                         const wxPoint& pos = wxDefaultPosition,
                         const wxSize& size = wxDefaultSize,
-                        long style = 0)
+                        unsigned int style = 0)
     {
         Create(parent, id, pos, size, style);
     }
@@ -82,7 +82,7 @@ public:
                 wxWindowID id = wxID_CONTEXT_HELP,
                 const wxPoint& pos = wxDefaultPosition,
                 const wxSize& size = wxDefaultSize,
-                long style = 0);
+                unsigned int style = 0);
 
 
     void OnContextHelp(wxCommandEvent& event);
