+++ include/wx/dfb/nonownedwnd.h
@@ -37,7 +37,7 @@ public:
                      wxWindowID id,
                      const wxPoint& pos = wxDefaultPosition,
                      const wxSize& size = wxDefaultSize,
-                     long style = 0,
+                     unsigned int style = 0,
                      const wxString& name = wxASCII_STR(wxPanelNameStr))
     {
         Init();
@@ -49,7 +49,7 @@ public:
                 wxWindowID id,
                 const wxPoint& pos = wxDefaultPosition,
                 const wxSize& size = wxDefaultSize,
-                long style = 0,
+                unsigned int style = 0,
                 const wxString& name = wxASCII_STR(wxPanelNameStr));
 
     ~wxNonOwnedWindow();
