+++ include/wx/docmdi.h
@@ -41,7 +41,7 @@ public:
                         const std::string& title,
                         const wxPoint& pos = wxDefaultPosition,
                         const wxSize& size = wxDefaultSize,
-                        long style = wxDEFAULT_FRAME_STYLE,
+                        unsigned int style = wxDEFAULT_FRAME_STYLE,
                         const std::string& name = wxFrameNameStr)
         : wxDocMDIParentFrameBase(manager,
                                   parent, id, title, pos, size, style, name)
@@ -75,7 +75,7 @@ public:
                        const std::string& title,
                        const wxPoint& pos = wxDefaultPosition,
                        const wxSize& size = wxDefaultSize,
-                       long style = wxDEFAULT_FRAME_STYLE,
+                       unsigned int style = wxDEFAULT_FRAME_STYLE,
                        const std::string& name = wxFrameNameStr)
         : wxDocMDIChildFrameBase(doc, view,
                                  parent, id, title, pos, size, style, name)
