+++ include/wx/gauge.h
@@ -71,7 +71,7 @@ public:
                 int range,
                 const wxPoint& pos = wxDefaultPosition,
                 const wxSize& size = wxDefaultSize,
-                long style = wxGA_HORIZONTAL,
+                unsigned int style = wxGA_HORIZONTAL,
                 const wxValidator& validator = wxDefaultValidator,
                 const std::string& name = wxGaugeNameStr);
